requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
gast==0.2.2
gluon2keras==0.0.2
gluon2pytorch==0.0.6
gluoncv==0.6.0
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.8.4
grpcio==1.27.2
h5py==2.10.0
idna==2.9
intel-tensorflow==2.1.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.2.0
mmdnn==0.2.5
mxnet==1.6.0
numpy==1.18.1
oauthlib==3.1.0
onnx==1.3.0
onnx2keras==0.0.18
onnxruntime==1.1.2
opt-einsum==3.2.0
Pillow==7.0.0
portalocker==1.5.2
prompt-toolkit==1.0.14
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
PyInquirer==1.0.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytorch2keras==0.2.3
PyYAML==5.3
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-addons==0.8.3
tensorflow-cpu==2.1.0
tensorflow-estimator==2.1.0
tensorflow-hub==0.7.0
tensorflowjs==1.6.0
termcolor==1.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.43.0
typeguard==2.7.1
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.12.0
